// src/app/shared/models/userRegisterRequest.ts
/**
 * BicycleCompany Authentication Microservice
 * A simple ASP.NET Core Web API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UserRegisterRequest { 
    email: string;
    firstName: string;
    secondName: string;
    password: string;
    passwordConfirm: string;
}